Modernize codebase and update to C++20/toolset v145
- Add default member initializers and explicit constructors for safer object initialization
- Replace fallthrough comments with [[fallthrough]] for C++20 compliance
- Suppress C4191 warnings in MFC message maps using pragmas
- Remove legacy build configs; update SDK to 10.0.26100.0 and toolset to v145
- Set language standard to C++20 and enable all warnings
- Delete obsolete binaries, help files, and intermediate artifacts
- Improve type safety, Unicode usage, and string parsing robustness
- Clean up and clarify code style, comments, and project/solution mappings

+++ AeSys/PrimState.cpp
@@ -17,7 +17,7 @@ const CPrimState& CPrimState::operator=(const CPrimState& src) {
   return (*this);
 }
 void CPrimState::Restore(CDC* deviceContext, int iSaveId) {
-  if (iSaveId >= sizeof(psSav) / sizeof(psSav[0])) return;
+  if (iSaveId >= static_cast<int>(sizeof(psSav) / sizeof(psSav[0]))) return;
 
   if (psSav[iSaveId] != 0) {
     SetPen(NULL, deviceContext, psSav[iSaveId]->PenColor(), psSav[iSaveId]->LineType());
